Define a reference to the root context as 'rc'
This allows us to hook up new rvalues via term rewriting, for example in the
Knockoff compiler. An example for such an rvalue is

data-bind="text: $i18n('login', {user: user, error: error})"

, which can compile to rc.g.i18n(..) in tassembly. The i18n function can
be registered on the tassembly instance's globals member, which is referenced
by the root context in the 'g' member.