Optimize raw mode by optionally assigning to val inline
So far, we have wrapped complex expression evaluation in a self-evaluating
function containing a try/catch block. This makes it an expression, which
makes it easy to work with it. However, the combination of the nested function
and try/catch is about three times slower than an inline try/catch.
Fortunately, the value of the expression is often assigned to a variable
`var`. This patch adds a parameter for evalExprStub that lets us indicate when
this is the case, and then compiles the expression into an inline `try { var =
expr; } catch (e) { .. }` block.